- Fixed Visual C++ and PSP build errors. - Added double buffering to Adlib and PC speaker sound for a speedup. - Changed the Sound Source emulation to use the new double buffering method. - Fixed Adlib rendering and PCM compatibility. - Added FIFO buffer 8&16 bit move functionality (read X data from one buffer into another when enough is gathered).

+++ emu/gpu/gpu_framerate.c
@@ -140,7 +140,7 @@ void renderFramerate()
 					if (framerateupdated) //We're updated!
 					{
 						framerateupdated = 0; //Not anymore!
-						CPUspeed = (byte)roundf(SAFEDIV(last_timing, (double)getnspassed_k(&CPU_timing))*100.0f); //Current CPU speed percentage (how much the current time is compared to required time)!
+						CPUspeed = (byte)round(SAFEDIV(last_timing, (double)getnspassed_k(&CPU_timing))*100.0f); //Current CPU speed percentage (how much the current time is compared to required time)!
 					}
 					GPU_textprintf(frameratesurface, RGB(0xFF, 0xFF, 0xFF), RGB(0xBB, 0x00, 0x00), "\nCPU speed: %i%%", CPUspeed); //Current CPU speed percentage!
 				}
@@ -153,7 +153,7 @@ void renderFramerate()
 				if (framerateupdated) //We're to be updated with the framerate rate!
 				{
 					framerateupdated = 0; //Not anymore!
-					CPUspeed = (byte)roundf(SAFEDIV(last_timing, (double)getnspassed_k(&CPU_timing))*100.0f); //Current CPU speed percentage (how much the current time is compared to required time)!
+					CPUspeed = (byte)round(SAFEDIV(last_timing, (double)getnspassed_k(&CPU_timing))*100.0f); //Current CPU speed percentage (how much the current time is compared to required time)!
 				}
 				GPU_textgotoxy(frameratesurface, 0, 0); //For output!
 				GPU_textprintf(frameratesurface, RGB(0xFF, 0xFF, 0xFF), RGB(0xBB, 0x00, 0x00), "CPU speed: %i%%", CPUspeed); //Current CPU speed percentage!
